PlaylistEditorView: don't use itemVisible if using the tree view
This will make sure we don't hunt for node images and icons if we are using the
tree view which doesn't show them.

It would be nice to only disable the hunt if the buttonimage image is not
present in the buttonlist but I can't see a way to do that.