Fix settings display on macs
Qt widgets uses a default style that varies between platforms. On Linux, the default text is white, on mac its black.
The color of the text in the old settings aren’t configurable as it is (wouldn’t be difficult to do so)

Most themes are using a dark background, so on mac, using the old type settings is proving difficult as mostly unreadable.

So we redefine the style of most QWidget objects so it looks more similar to Linux.
This is done using Qt stylesheets. Where Qt stylesheets do not provided the intended behavior, we use QPalettes instead