simplify handling of missing endtime at end of file in mythfilldatabase
Instead of making up endtime of "next midnight" and "next 06:00 am" we
just let the database fill in its default of 0000-00-00 00:00 and replace
it later with whatever the next starttime is on that channel.

This has seen very light testing, so I'm commiting it to master only. If
I get positive reports from users I'm happing to merge it into fixes.

Fixes #11541