Fix old Qt setting screens preventing them from accepting keyboard inputs
Hopefully this is the final fix…
The issue seems to only have occurred on macs. When opening two MythDialogs (that includes MythPopup) in a row, often the 2nd one would not accept any used input anymore, only causing a “beep”; making it impossible to even exit that window.

The existing logic (and there was nothing wrong with it) was to simply disable the current MythDialog and put the new one on top of it. However, when doing so, somehow key events are now ignored.
So instead, we now disable all widgets within that MythDialog, without disabling the dialog itself.

When doing so, the newly created MythDialog is always working as expected.
This code should not be necessary, it is a workaround to what appears to be a Qt bug.

Fixes #11990.