Improve readability of old settings on mac
Text Line background was the same color as the selection. So it was impossible to tell if some text was selected or not. Set background of editable text to a very light blue and keep the text black.
That selection color issue also exists on linux, so we could take the same approach there