Fix settings display on macs
My analysis for the problem in 96ecd539de8e76a880f0013b4b8854da216e17b1 was wrong.
The actual issue is that QPalette inheritance appears broken on a mac; if you set the palette to the top widget container, it doesn’t do anything on its children. However, setting the palette to the main application itself does.

Now things look exactly the same between mac and linux for better or worse. Including font sizes and colors.
A downside, is that the selection color appears to be misused in most themes including the default. So you can't tell if a text is selected or simply having focus

(cherry picked from commit 15425cb53e643a8c00975c1ef0ea7674464fdca2)