Add ReadInternalMode to RingBuffer class
This mode allows to limit the content of the ringbuffer to the content already loaded. Seeking, reading and other operations will not be destructive.
This allows for almost instantaneous streams and bitrate detection.

Duration calculated will obviously be wrong, but as the duration is stored in the DB we’ll use that duration instead anyway, meaning we only have to worry about streams detection itself.

This greatly speed up liveTV transition change and playback startup