Revert "Fix GuideGrid to include rsOtherShowing in the 'will record' highlighting colour. Fixes #12081"
This reverts commit 82b02823c4235b21f40a3722cef02e2387722f8f.

Until I re-write the frontend guide grid to match the WebFrontend
guide, this has the potential to cause more confusion that it would
solve. In the WebFrontend guide earlier/later/other showings get
highlighted in a completely different colour making it clear what will
be recorded, what won't be recorded (old episodes/duplicates) and what
will be recorded - just not at that time or on that channel.

Conflicts:
	mythtv/programs/mythfrontend/guidegrid.cpp