MythNews: Fix parsing of the enclosure on some feeds
There are some broken feeds that don't add a proper type for the enclosure
which was causing us to assume it was a video and try to play it even if it
was an image.

So if no type is given check the filename extension of the url for any image
extensions before assuming it must be a video.

(cherry picked from commit 7ee573ae67992c2f4c34c300740be713f82d1e98)