MythMusic: Release audio decoder when music paused and tv playback wanted.
When "continue play music" is active and within the MiniPlayer pause is
called, a tv/video playback can not open the audio device.

In MusicPlayer::StopPlayback the check m_isPlaying is false, therefor not
stopping the decoder.

Fix: Instead of checking m_isPlaying, use m_output the check for an existing decoder.

#Fixes #12063.

Signed-off-by: Paul Harrison <pharrison@mythtv.org>