Support limited multirec on digital tuners that don't use multiplexes.
This is primarily intended for cable card tuners like the Ceton and
HDHomerun Prime.  In theory, other tuners which produce digital
output, such as firewire and the analog side of the HVR-2250, could do
this too when they are converted to use the newer stream handler.

This change will not magically make these cards capable of recording
more than one channel at a time on the same tuner.  It will, hoever,
let the same content be written to different recordings when they
overlap because of pre/post-roll or start-early/end-late settings.

Note that the scheduler does not currently try to place overlapping
programs to take advantage of multirec.  A future change will try to
address that.