Keep some recordings from getting a LiveTV recgroup
Following #10943, an ApplyRecordRecGroupChange() was added near the
end of GetProgramRingBufferForLiveTV() to update recorded.recgroup
before it is queried in StartedRecording().

The problem arises because we don't have the correct recstartts for
ApplyRecordRecGroupChange() until later when we successfully insert
the program in RecordingInfo::StartedRecording.

Fixes #11119
Signed-off-by: David Engel <dengel@mythtv.org>

(cherry picked from commit 9c666b602edc41781853e0698191f3239341a33b)