Fix rewind during liveTV or in progress recordings
Similar issues that cause FastForwarding to fail (fixed in 981b45871ab0faefa03f5f0e8a2a278f2a4d1761)
Under some circumstances, rewind would jump to a much earlier time than requested.
Issue is that the size of the file was obsolete and we would jump to the last known size.

Now we make sure we get an updated value from the backend when necessary.

Also, a small optimisation in FastForward, we jump to -3s in liveTV/In progress before the end at the most. Prevent stuttering