Don't unnecessarily retune Ceton and HDHomerun Prime when using multirec.
Previously, when using multirec on a Ceton or HDHRP tuner and a second
recording started, the channel would be retuned and cause a small
disruption in the first recording.  This change doesn't perform the
useless retuning avoiding the disuption on the first recording.

This fix was originally developed by John Poet with a couple of fixes
by me after testing.