WebFrontend: Make the guide timebar static.
This one tiny detail has caused me more headaches than almost any
other bit of the WebFrontend so far. I looked at several approaches
and hit many dead ends. This _should_ be pure css, but it doesn't
appear possible. I investigated giving the timebar fixed positioning,
but that requires too many compromises, too many wrappers (css was
supposed to do away with markup for the purpose of styling) and it
would make updates to the page harder.

A fancy javascript (jQuery) carousel type approach proved to be way
too slow on mobile devices. Javascript just is never going to be fast
enough on phones/tablets.

So going back to a simple approach I've mixed css (overflow: scroll)
with a one liner of javascript to resize the div to the remaining
screen height (impossible in css). However there is still an
irritating mis-alignment issue on mobile browsers because they don't
display the scrollbar, this results in the timebar being slightly
mis-aligned with respect to the programme boundaries. Since I've
already wasted a lot of time on this, I'm committing it as is and will
revisit the alignment problem if/when time allows.