Fix connection to Master Backend where the master IP address is a link-local address
Similar problems as fixed in 5594a6f32e73b0519b556b63983786a01a0f2228, if MasterServerIP contains a scope Id, it will prevent frontend and slave backend to connect.

Also, the master backend detects that it is the master by comparing its IP address with the value stored in the MasterServerIP setting. The comparison is made using strings.
So if MasterServerIP is set to something like 192.168.000.001 it will fail (as we would be comparing with 192.168.0.1). IPv6 can also contains upper or lower case characters.

Add convenience methods GetMasterServerIP, GetMasterServerPort and GetMasterServerStatusPort ; which will ensure the value read is always correct.

Ref #10921