Added support for saving the recording input name in the recorded table.
This makes it possible to identify on which tuner a recording was made
after it was recorded.  This can be very useful in diagnosing
recurring problems with a tuner that aren't noticed right away.  The
input name is used rather than the input number because it is expected
to be more user friendly and useful, especially if the tuners have
been renumbered since the recording was made.

This is based on patches submitted by George Nassas.

Refs #11954