Allow MasterServerIP, BackendServerIP and BackendServerIP6 to be a hostname
This became a requirement following 5594a6f32e73b0519b556b63983786a01a0f2228.
It was assumed in that commit that MasterServerIP was always an IP address as it’s the only thing mythtv-setup will allow.
However, pre-0.25 it was possible to set a hostname. Breaking grand-fathered network configuration

Add MythCoreContext::resolveSettingAddress and resolveAddress utility method to resolve hostnames