Services API: Add Details arg for GetChannelInfo endpoint.
Defaults to false.

Fetching some of this extra info means an extra query for each channel
making it slow and I'm struggling to imagine a scenario where a
Services client would need the frequency/modulation info for a
channel. It's a fairly useless subset of the information available by
calling GetVideoMultiplex() so don't be surprised if this information
disappears from the ChannelInfo object in future.