Bump Database Query logging to LOG_INFO|VB_DATABASE.
The database verbose level is pretty useless without being able to see the query logging and including them with --loglevel debug is problematic because we also include the results and we don't have to wade through tens of thousands of results to inspect the query strings. You can always reproduce the result if you know the query used, but you can't reproduce the query from the result.