Services API: Eliminate overlap between ChannelInfo and VideoMultiplex.
This bumps the ChannelInfo object version to 2.0.

If you were previously using this information, you will need to call
the GetVideoMultiplex endpoint with the mplexid from the ChannelInfo
instead.

Fetching this internal tuning information for every single channel was
very slow and yet most users of the Services API will never need access to
it. In fact its of arguably of no use to anyone using the API at all.