Add MythUSWCCopy YV12/SV12 conversion utility class
This commit adds Uncacheable Speculative Write Combining memory optimised methods to copy and convert frames.
USWC are typically used by hardware decoder, in particular it is use by VAAPI and DXVA2.
Using this method yield significant speed increase when working with USWC memory; however it is significantly slower with standard, cached memory.

So we add a crude detection method by simply checking on which one is the fastest and use the appropriate method for future calls.

More information is available at:
https://software.intel.com/en-us/articles/copying-accelerated-video-decode-frame-buffers