Try to prevent duplicate recording rules from getting created.
Add a unique index on the record table to prevent the creation of
duplicate recording rules, where duplicate for this means having the
same chanid, starttime/date, title and rule type.  This is mainly
intended for the Web GUI where an accidental double click could result
in truly duplicate rules.