UPnP - Fix broken FFWD/REW/JUMP in playback for LG SmartTV DLNA clients for recordings with startoffset
This fixes error with playback navigation through FFWD/REW buttons and navigation bar in some LG SmartTV DLNA clients. The patch replaces dtProgStart by dtStartTime in the query to recordedmarkup table.

If startoffset for a recording is nonzero, dtProgStart by dtStartTime differ and the query fails. As fallback, the duration attribute for <res/> tag is filled by the value from EIT data, which may differ from the actual length of playback. The DLNA client concerned unfortunately relies on duration attribute for playback navigation, so the playback navigation is completely broken.