Fix PiP with VDPAU on mesa's drivers
This is due to a bug in mesa's vdpau implementation.
https://bugs.freedesktop.org/show_bug.cgi?id=80561

So as a work around, create a dummy bitmap surface and use this.
Limit the work around to non-nvidia drivers

There is still a remaining issue, only one PiP shows at a time. This is due to another bug in mesa's vdpau VdpVideoMixerRender, it clears the whole destination surface. So you only see the last PiP window.