MythUI: Add depends support for buttonlists.
This lets you show text, images or groups of widgets only when there
are items in the named list, or alternatively hide them if there are
no items.

Possible uses include messages explaining that there are no items:
 e.g. 'You have no recordings available'
Or images behind or around lists that disappear when the list is empty.

This will give themers more flexibility in their presentation. Icons
could be used instead or alongside text for existing "no items" type
messages.