mythutil: add an option to check recordings
This checks the integrity of all recordings and prints a summary of what it
finds. Currently it checks if the file exists, they are not zero bytes in size
and they have a seek table. Other test may be added later.

You can optionally fix the seek tables of any recordings were it is missing.

You run the command like this:

mythutil -v none --checkrecordings --fixseektable