Specify supported previous grabbers in tmdb3.py
This specifies that tmdb3.py is able to accept the inetref indexing used
by past tmdb.py and tmdb.pl. The metadata library will then use tmdb3.py
for any existing content tagged with "tmdb.py:xxxx" or "tmdb.pl:xxxx".

Those grabbers were deprecated before the inetref tagging was added, and
so will never be used in practice, but are added for example of how to
use the mechanism.

Refs #12149