metadata: uses _ as separator in inetref in place of :
the inetref is used in the artwork filename, but the : character isn’t permissible with all file systems (FATx, VFAT, exFAT, HFS, and AmigaOS in particular).
the : used in old inetref is kept for backward compatibility.

Attempt was made to generate a filename not using the inetref as-is, but the change was very disruptive as there are about 12 functions across various classes that generate those filenames. Likelihood of missing a case was too great.

(cherry picked from commit 6fddce08ff196a6b78b066dede0558593ba86f80)