MythGallery: optionally allow import scripts to be run
This allows you to run a script to import images which is useful if your camera
doesn't provide a mountable directory of images to import but requires a
utility like gphoto2 to download the images.

By default allowing the import to run a script is disabled for security reasons
but can be be enabled by checking the 'Allow the import to run external script'
setting.

You set the script to use in the 'Paths to import images from' setting which
will accept both a path and a script. The script should accept a single
parameter which is the directory to save the images to.

An example script for qphoto2 would be :-

  #!/bin/bash
  #$1 is the directory to save the images to

  cd $1
  gphoto2 --get-all-files