metadata: complete refactor of search
Use different sequence of searches in order to find the best matching possibility.
For TV, should no match be found, search for a generic artwork. This generic artwork will be automatically upgraded should specialised artwork ever be found.
This is done by setting a season number 0 for that generic artwork.
Original season/episode number will never get rewritten like it used it (was set to 1x1)