SendReceiveStringList: handle the UNKNOWN_COMMAND response properly
Change it to return false and log the error rather than return true
and carry on as though nothing is wrong.

This should never happen in reality but if it does SendReceiveStringList
should not return true and carry on as normal since we don't handle that
well in a lot of places.