Fix null pointer derefence in protocol error handling. Coverity 1228785
The PlaybackSock object has been deleted at this point and the pbs
pointer is NULL. This would have triggered a backend segfault if a
malformed MYTH_PROTO_VERSION query was received.

Instead use a pointer to the underlying socket (MythSocket)
which is what we really want ultimately. This adds a second
SendErrorResponse which handles MythSocket, so that either a
PlaybackSock or MythSocket pointer may be used.