Initialize cleartimeonpause to false in NuppelVideoRecorder. Coverity 1025911
This was a genuine bug. It's not clear what the correct initial value
should have been. This code is 11 years old, no-one who wrote it is
around anymore. Since the bug is so old and the analogue frame grabber
support may be removed soon there is no reason to worry about it.