Allow initial ProgLister item to be set by time.
This is primarily intended for bringing up the Channel Search screen
from another screen.  Instead of setting the first item listed as
current, it tries to start with the same item that was current on the
other screen.

Note that the value used to save the desired time is cleared upon
first use.  If there is a desire to keep it current and use it when
changing views, like chaning channels on the Channel Search screen,
some more work will be needed.