Fix more IP vs hostname confusion in myth:// urls
A myth url consists of <group>@<hostname>:<port> where <hostname> is
the unique identifier of a backend, NOT an IP address. Some code
rightly expects to be able to extract the hostname from the URL and
use it as the unique identifier, for example to StorageGroup() or as a
key for DB lookups. The places where an IP was used broke this
assumption and caused the code to take an extended route to complete
it's task, or caused failures.

The 'hostname' will be converted to an IP address only at the point
where we connect to the remote backend to transfer/delete/etc the
file. Since only mythtv is expected to handle myth:// urls this is OK.

Since this has been causing confusion for some time there may be parts
of the code that incorrectly expect an IP address which will have to
be fixed.