Migrate information from recorded and recordedmarkup to recordedfile.
The old data is left in place for now, as code still needs to be
changed to use the new table.

Codec and bitrate info is missing for older recordings. One possible
solution to this will be an external utility which inspects all
recordings and updates the database. Possibly as a one-off event when
upgrading from 0.27 to 0.28, or as part of the daily housekeeper.