WebFrontend: More animation performance improvements, and other tweaking
Instead of animating through JS (jQuery), use native CSS3 transitions and
animations which are faster because they are handled in native code
and hardware accelerated. It's also the Right Way™ to do visual stuff
in web pages and makes it easier to tweak effects for different screen
sizes and platforms.