MythZoneMinder: Add a popup mini player to show live camera views
This adds a jump point which will show a mini player that shows live views
from the ZoneMinders cameras.

The mini player works in a similar way to the music miniplayer so it will show
for 10 seconds then auto hide itself unless you press SELECT which will stop
the popup auto hiding. You can close the popup window at any time with ESCAPE.
The 1 key will cycle through the available cameras same as the normal live
viewer.

The next step is to auto show the mini player when an alarm is triggered in
ZoneMinder.