MythZoneMinder: Add automatic popup notifications on alarms
Whenever ZoneMinder triggers a motion alarm for a monitor (camera) the popup
mini player will show with the alarmed monitor selected.

Only monitors that have notifications enabled will trigger the mini player to
show. You can toggle which monitors show notifications in the MythZoneMinder
console screen press MENU to show the options popup for the highlighted
monitor and make sure 'Show Notifications' is checked.

One use case would be a front door camera that triggers the popup to show
when someone is at the door so you know who it is before you answer it.