Temporarily disable editing of cards, inputs and sources.
This is in preparation for merging cards and inputs.  Doing so allows
the succeeding changes to be done in smaller, logical steps without
breaking core functionality.  The latter could be important if this
series of changes are part of a git bisect.  Keeping the edit
funcionality active would require making a lot of related changes
altogether in one, very large commit.