Split multi-input cards and renumber inputs to match cards.
For each second or higher input on a card, create a new card with the
same information and attach the input to the new card.  After this, a
card can have at most one configured input.  Next, renumber the inputs
such that the inputid is equal to the cardid.  Finally, automatically
create an inputgroup for each unique host/device and attach each
card/input to the appropriate inputgroup.

Note that this change subtly introduces the concept of device and user
inputgroups.  The former are automatically managed when cards are
edited and supports conflict detection for cards using the same
device.  The latter are for the user to specify cross-device
relationships that cannot be automatically detected.  This commit only
contains the minimum changes needed to implement the input splitting
and renumbering.  The rest will come later when card/input editging is
re-enabled.