Frontend Network Control: add commands to get/set the area of a widget
This adds two new commands to get or set the area of a widget on the active
screen.

The get command is:
theme getarea WIDGET

For example to get the area of the clock widget:-
# theme getarea clock
The area of 'clock' is x:780, y:40, w:480, h:40

The set command is:
theme setarea WIDGET X Y WIDTH HEIGHT

For example to set the area of 'clock':-
# theme setarea clock 780 80 480 40
Changed area of 'clock' to x:780, y:80, w:480, h:40

This is just a proof of concept to test the idea of a theme editor that uses
the network control to control a frontend to display a live WYSIWYG view of
the theme while you make changes.