Replace uses of cardinput table with capturecard table.
Move the contents of the cardinput table to the capturecard table and
update all database queries accordingly.  cardid is now synonymous
with cardinputid.  There are many, many, many places in the code where
the logic now makes little to no sense.  For the most part, those have
been left unchanged and will be addressed later.