Update and re-enable editing of capture cards.
Fix editing of capture cards after the merge of the capturecard and
cardinput tables.  A "Device" inputgroup is now automatically created
for each unique hostname/videodevice.  This is now how virtual tuner
and multiple input cards are tied together -- cards using the same
device are automatically assigned to the same inputgroup.

Note, it turns out that there were several existing bugs that show up
when the type or videodevice of an existing card is changed.  This
change does not do much to address this problem, so don't do that.  It
will hopefully get addressed some time in the future, probably when
the editing of inputs is merged with card editing.