Null pointer check in MainServer::customEvent.
There are a few backtraces from ubuntu users appearing to show a crash here with
a null event pointer. It's not possible to work backwards and see how a null
event managed to end up being posted in the first place, and this
change won't necessarily prevent all variations of this crash but it's
the best I can do for now without a full audit of all code using the
events system.