WebFrontend: Add a restrictive Content-Security-Policy header
This adds an additional layer of protection to the WebFrontend against
most forms of XSS exploit by blocking ALL external content - scripts,
css, images, media etc.

Such a restrictive policy is fine for the WebFrontend as by design it
has to work on isolated networks, or even single machines with no
internet access.

For now, XHR connections to external hosts are still permitted as
that's the mechanism by which we trigger playback of a recording on a
remote frontend. This may be switched to an internal call at some point.

I've also moved setting of some response headers into the HTML server
extension as they aren't relevant to the wider HTTP server.