Update and re-enable editing of card inputs.
Fix editing of card inputs after the merge of the capturecard and
cardinput tables.  The inputgroups that can be configured here are now
known as "User" inputgroups.  These allow the user to define input
relationships that cannot be automatically determined.

Similarly with the capture card editing, strange things can happen
when multiple inputs are configured with the same device and
inputname.  Don't do that until a better design can be implemented.