WebFrontend: Show current frontend state info on the overview page.
This is a proof of concept and nothing more, it's not really practical
to handle this all client side as we do now since as there are potential
mixed-content and security issues. The backend might have ssl enabled,
but not the frontend for example. The frontend port info is hardcoded,
when it should instead come from the settings via the GetFrontends endpoint.

The existing format of the frontend state info needs work too, it's
too messy to parse and violates some of the principals of good xml. We
need the backend to provide state info as a translated string for the
UI.

Expanding it to display what is being watched and even the current
position are ultimate goals.